리뷰작성가능 리뷰 리스트, 작성한 리뷰 리스트

+++ src/main/java/com/spring/green2209s_08/web/domain/Review.java
@@ -12,13 +12,14 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Review extends CommonEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+
+    @Id  @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "review_id")
     private Long id;
+
     private String headLine;
     private String content;
-    private int reviewRating;
+    private Long reviewRating;
 
     @ManyToOne
     @JoinColumn(name="item_id")
